[FIX] stock: assign serial button
We display the button on the move if detailed operations is enabled and
if pre-fill detailed operations is disabled. There was a mix of computed
field/conditions in the xml resulting in an always hidden button.

We now only use the computed field.

related to db12a0199ef86dce143a1e4b12ec531c28099c2a

closes odoo/odoo#36519

Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>