[FIX] sale_timesheet: display timesheeted quantity correctly on sale order
There was a stat-button on sale orders coming from a task with
timesheet that displayed the quantity of timesheets linked to the
origin task. This stat-button was re labeled with the rimesheet uom
(hours/days/...) but the field still displayed the number of
timesheets instead of the total timesheeted amount in the correct
uom.

This commit introduces a new computed field on the sale order, which
computes the amount of time on timesheets depending on the used uom.

The compute method uses the product_uom_hour record as it is the uom
in which the amount is stored on the model, that record is guaranteed
to exist as it's model unlink method was overridden to prevent
removing working-time uom.

Task-2065079