[FIX] web: Display archive action if active in view
Since 35722a8, "Achive" and "Unarchive" actions are
automatically displayed in the actions dropdown
if the record has an `active` field.
The "Archive" action is shown if `active=True`, and
"Unarchive" otherwise.

However, if the record has the `active` field but
it's not in the view, the form controller think
`active=False` because `undefined` is falsy.
Therefore, the "Unarchive" action is always shown.

If `active` is not in the view, neither "Archive" nor
"Unarchive" should be shown.

closes odoo/odoo#36551

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>