[FIX] industry_fsm: fix sale order issue when adding product without timesheet on tasks
The link between a sale_order and a task relied on a sale_line between the sale order and the tasks. This was working when allow timesheet and allow material were true, as we had a timesheet_product_id and the sale_line was created empty. But when we create a project with material but no timesheet, the sale line would crash.

Now the sale order and the sale line creation are split, and we write the sale order id on the tasks. This works because with the new ORM, we can write on a stored compute field, and if we take care of not overwriting it in the compute, it stays at what we wrote.

task-2047389		closes-5388