[FIX] base: address and name split on 2 lines
The name and address printed on any PDF report (e.g. SO or RFQ) is very
likely to be split on 2 lines once it is long enough. For example,
'My Company (Chicago)' will be displayed on 2 lines.

This is due to c701fb62a02db5 and 18cbfc89ee8b0d0.

opw-2061553

closes odoo/odoo#36498

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>