[FIX] web: unable to scroll modal's table in Safari (iOS)
Before this commit, users wasn't able to scroll vertically
when there was "table-responsive" class on modal's table.
In fact the scroll was effective on the body and not
in the modal content.

We can fix this issue by having flex elements but as this is
a fix, we preferred to choose a one line fix solution.

Steps to reproduce :
* open Odoo in Safari (iOS)
* open Sale app
* create a new quotation
* select a customer
* select "Search More ..."
* try to scroll the table vertically (BUG)

Sources :
https://stackoverflow.com/questions/41681251/overflow-x-scroll-inside-a-fixed-div-prevents-vertical-scrolling-on-ios#comment84970666_41693350
https://stackoverflow.com/a/57566116

opw-2042095

closes odoo/odoo#36707

closes odoo/odoo#36730

Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>