[FIX] point_of_sale: prevent opening session if no chart of accounts
Before this commit, a user can create a Point of Sale record and
proceed on launching a new session. If the user ignored the warning
message regarding the installation of Chart of Accounts, the opened
session will not work properly and the session can't be closed.
This commit prevents the user from launching a new session if there is
no Chart of Accounts installed for the current company.

closes odoo/odoo#36777

Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>