[FIX] purchase_requisition: supplier info
When a blanket order is opened, a supplier info is created and should be
used in priority over a regular supplier info with the same values.

That's what rev [0] tried to do but there was some mistakes: the
`purchase_requisition_id` field added in the order is ignored since it
isn't stored and the test didn't failed because the created blanket has
a lower price unit than the regular supplier info, meaning it is used in
priority. Setting the same price unit (50) result in the new purchase
order line merged in the first one.

Do not re-implement the same order.
Adapt the test now that we know the order didn't work (see rev[1]). The
user has to manually set a higher priority to the supplier info
generated by the purchase requisition.

[0] acb0df1a24266150b4bc3fa337709831eeddab26
[1] 387b1be0490f97570382bd897fb780f091ed1626

closes odoo/odoo#36840

Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>