[FIX] web_editor, *: restore crop dialog
* website

Following the new editor merge at https://github.com/odoo/odoo/pull/29775
(which was finally partially reverted for 13.0), the crop dialog was not
working anymore on sub-sequent crop. This was due to 2 things:

1) Re-cropping an image that was just cropped now

The original source of the recently cropped image is saved as temporary
jQuery data on the image element. Unfortunately, for some strange
reason, that image is cloned before any edition and thus lose that data.
This commit fixes the problem by cloning the data too... supposing the
orignal clone is there for a reason.

2) Re-cropping an image that was saved before

This did not work anymore simply because... the cropped images were not
saved in database anymore but saved as base64 data in the view... For
some reason, the cropped images to save were searched in the original
DOM instead of the edited DOM...

Also fixing the random use of a mutex.

Discovered while working on task-2059480

closes odoo/odoo#36878

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>