[IMP] http: rollback from checked_call only if necessary
The rollback clears the cache, which lose all data that have been fetched before
arriving in the route method.

This lost cache includes some website data that was used during the dispatch and
that will be used again in the route.

By keeping it we reduce the number of queries on every request by at least 2.

Part of task-2061122

closes odoo/odoo#36245

Signed-off-by: Raphael Collet (rco) <rco@openerp.com>